requirements.txt
asyncpg
aiogram
aiohttp
aioredis
python-dotenv
gino
requests
qiwipy
loguru
openpyxl